refactor: Complete CLI refactoring cleanup
- Replace original CLI files with refactored versions
- Remove _refactored suffix from all files
- Update all imports to use the new module names
- Maintain backwards compatibility with main() function
- All tests passing with the consolidated structure

This completes the CLI refactoring effort, with all functionality
now properly separated into validators, formatters, and handlers.
