+++ src/scriptrag/cli/commands/__init__.py
@@ -6,13 +6,13 @@ from scriptrag.cli.commands.init import init_command
 from scriptrag.cli.commands.list import list_command
 from scriptrag.cli.commands.mcp import mcp_command
 from scriptrag.cli.commands.pull import pull_command
-from scriptrag.cli.commands.query import get_query_app
+from scriptrag.cli.commands.query import create_query_app
 from scriptrag.cli.commands.search import search_command
 from scriptrag.cli.commands.watch import watch_command
 
 __all__ = [
     "analyze_command",
-    "get_query_app",
+    "create_query_app",
     "index_command",
     "init_command",
     "list_command",
