+++ src/scriptrag/cli/commands/scene.py
@@ -1,10 +1,7 @@
-"""Scene management commands for AI-friendly editing."""
+"""Refactored scene management commands with clean separation of concerns."""
 
-import asyncio
-import sys
-from datetime import datetime
 from pathlib import Path
-from typing import Annotated, Literal
+from typing import Annotated
 
 import typer
 from rich.console import Console
@@ -12,15 +9,19 @@ from rich.console import Console
 from scriptrag.api.scene_management import SceneManagementAPI
 from scriptrag.api.scene_models import SceneIdentifier
 from scriptrag.cli.commands.scene_config import load_config_with_validation
-from scriptrag.cli.scene_formatter import SceneFormatter
+from scriptrag.cli.formatters.base import OutputFormat
+from scriptrag.cli.formatters.scene_formatter import SceneFormatter
+from scriptrag.cli.utils.cli_handler import CLIHandler, async_cli_command
+from scriptrag.cli.validators.project_validator import ProjectValidator
+from scriptrag.cli.validators.scene_validator import (
+    SceneContentValidator,
+    ScenePositionValidator,
+    SceneValidator,
+)
 from scriptrag.config import get_logger
 
 logger = get_logger(__name__)
 console = Console()
-formatter = SceneFormatter(console)
-
-# Position type for scene operations
-PositionType = Literal["after", "before"]
 
 # Create scene subcommand app
 scene_app = typer.Typer(
@@ -32,7 +33,8 @@ scene_app = typer.Typer(
 
 
 @scene_app.command(name="read")
-def read_scene(
+@async_cli_command
+async def read_scene(
     project: Annotated[
         str, typer.Option("--project", "-p", help="Project/script name")
     ],
@@ -66,80 +68,65 @@ def read_scene(
     Examples:
         scriptrag scene read --project "breaking_bad" --season 1 --episode 1 --scene 3
         scriptrag scene read --project "inception" --scene 42
-        scriptrag scene read --project "inception" --bible  # List available bible files
+        scriptrag scene read --project "inception" --bible
         scriptrag scene read --project "inception" --bible-name "world_bible.md"
     """
-    try:
-        # Load settings with proper precedence
-        settings = load_config_with_validation(config)
+    # Initialize handlers
+    handler = CLIHandler(console)
+    formatter = SceneFormatter(console)
+    project_validator = ProjectValidator()
 
-        # Initialize API
-        api: SceneManagementAPI = SceneManagementAPI(settings=settings)
+    # Validate inputs
+    project = project_validator.validate(project)
 
-        # Check if reading bible content
-        if bible or bible_name is not None:
-            # Reading bible content
-            from scriptrag.api.scene_models import BibleReadResult
+    # Load settings
+    settings = load_config_with_validation(config)
 
-            bible_result: BibleReadResult = asyncio.run(
-                api.read_bible(project, bible_name)
-            )
+    # Initialize API
+    api = SceneManagementAPI(settings=settings)
 
-            if not bible_result.success:
-                console.print(f"[red]Error: {bible_result.error}[/red]")
-                raise typer.Exit(1)
+    # Determine output format
+    output_format = OutputFormat.JSON if json_output else OutputFormat.TEXT
 
-            # Use formatter to display bible content
-            formatter.format_bible_display(
-                bible_result.content,
-                bible_result.bible_files,
-                project,
-                bible_name,
-                json_output,
-            )
-            return
+    # Handle bible reading
+    if bible or bible_name is not None:
+        result = await api.read_bible(project, bible_name)
+        if not result.success:
+            handler.handle_error(Exception(result.error), json_output)
+        formatter.print(result, output_format)
+        return
 
-        # Reading scene content
-        if scene is None:
-            console.print(
-                "[red]Error: Either --scene or --bible must be specified[/red]"
-            )
-            raise typer.Exit(1)
-
-        # Create scene identifier
-        scene_id: SceneIdentifier = SceneIdentifier(
-            project=project,
-            scene_number=scene,
-            season=season,
-            episode=episode,
+    # Handle scene reading
+    if scene is None:
+        handler.handle_error(
+            ValueError("Either --scene or --bible must be specified"),
+            json_output,
         )
 
-        # Read scene (run async operation)
-        from scriptrag.api.scene_models import ReadSceneResult
+    # Validate and create scene identifier
+    scene_validator = SceneValidator()
+    scene_id = scene_validator.validate(
+        {
+            "project": project,
+            "scene_number": scene,
+            "season": season,
+            "episode": episode,
+        }
+    )
 
-        result: ReadSceneResult = asyncio.run(api.read_scene(scene_id))
+    # Read scene through API
+    scene_result = await api.read_scene(scene_id)
 
-        if not result.success:
-            console.print(f"[red]Error: {result.error}[/red]")
-            raise typer.Exit(1)
+    if not scene_result.success:
+        handler.handle_error(Exception(scene_result.error), json_output)
 
-        # Use formatter to display scene
-        if result.scene:
-            formatter.format_scene_display(
-                result.scene,
-                scene_id,
-                result.last_read,
-                json_output,
-            )
-
-    except Exception as e:
-        logger.error(f"Failed to read: {e}")
-        console.print(f"[red]Failed to read: {e}[/red]")
-        raise typer.Exit(1) from e
+    # Format and output result
+    formatter.print(scene_result, output_format)
 
 
 @scene_app.command(name="add")
-def add_scene(
+@async_cli_command
+async def add_scene(
     project: Annotated[
         str, typer.Option("--project", "-p", help="Project/script name")
     ],
@@ -159,6 +146,7 @@ def add_scene(
         str | None,
         typer.Option("--content", help="Scene content (or pipe from stdin)"),
     ] = None,
+    json_output: Annotated[bool, typer.Option("--json", help="Output as JSON")] = False,
     config: Annotated[
         Path | None,
         typer.Option(
@@ -168,114 +156,85 @@ def add_scene(
         ),
     ] = None,
 ) -> None:
-    r"""Add a new scene with automatic renumbering.
+    """Add a new scene with automatic renumbering.
 
     Scene content can be provided via --content or piped from stdin.
     Content must be valid Fountain format starting with a scene heading.
 
     Examples:
         scriptrag scene add --project "inception" --before-scene 10
-
-    Use --content flag or pipe content via stdin.
     """
-    try:
-        # Validate position arguments
-        if after_scene is None and before_scene is None:
-            console.print(
-                "[red]Error: Must specify either --after-scene or --before-scene[/red]"
-            )
-            raise typer.Exit(1)
+    # Initialize handlers
+    handler = CLIHandler(console)
 
-        if after_scene is not None and before_scene is not None:
-            console.print(
-                "[red]Error: Cannot specify both --after-scene and --before-scene[/red]"
-            )
-            raise typer.Exit(1)
+    # Validate inputs
+    project_validator = ProjectValidator()
+    project = project_validator.validate(project)
 
-        # Get content from stdin if not provided
-        if content is None:
-            if sys.stdin.isatty():
-                console.print(
-                    "[red]Error: No content provided. Use --content or "
-                    "pipe from stdin[/red]"
-                )
-                raise typer.Exit(1)
-            content = sys.stdin.read()
+    # Validate position
+    position_validator = ScenePositionValidator()
+    reference_scene, position = position_validator.validate(
+        {"after_scene": after_scene, "before_scene": before_scene}
+    )
 
-        # Determine reference scene and position
-        if after_scene is not None:
-            reference_scene: int = after_scene
-            position: PositionType = "after"
-        else:
-            # At this point, before_scene must be not None due to validation above
-            if before_scene is None:
-                raise ValueError(
-                    "Internal error: before_scene is None after validation"
-                )
-            reference_scene = before_scene
-            position = "before"
+    # Get content
+    if content is None:
+        content = handler.read_stdin(required=True)
 
-        # Create scene identifier for reference
-        scene_id: SceneIdentifier = SceneIdentifier(
-            project=project,
-            scene_number=reference_scene,
-            season=season,
-            episode=episode,
+    # Validate content
+    if content is not None:
+        content_validator = SceneContentValidator()
+        content = content_validator.validate(content)
+    else:
+        handler.handle_error(ValueError("No content provided"), json_output)
+        return
+
+    # Load settings
+    settings = load_config_with_validation(config)
+
+    # Initialize API
+    api = SceneManagementAPI(settings=settings)
+
+    # Create scene identifier for reference
+    reference_id = SceneIdentifier(
+        project=project,
+        scene_number=reference_scene,
+        season=season,
+        episode=episode,
+    )
+
+    # Add scene through API
+    result = await api.add_scene(reference_id, content, position)
+
+    if not result.success:
+        handler.handle_error(Exception(result.error), json_output)
+
+    # Format success message
+    if json_output:
+        handler.handle_success(
+            "Scene added successfully",
+            {
+                "created_scene": (
+                    {"number": result.created_scene.number}
+                    if result.created_scene
+                    else None
+                ),
+                "renumbered_scenes": result.renumbered_scenes,
+            },
+            json_output=True,
         )
-
-        # Load settings with proper precedence
-        settings = load_config_with_validation(config)
-
-        # Initialize API
-        api: SceneManagementAPI = SceneManagementAPI(settings=settings)
-
-        # Add scene
-        from scriptrag.api.scene_models import AddSceneResult
-
-        result: AddSceneResult = asyncio.run(api.add_scene(scene_id, content, position))
-
-        if not result.success:
-            console.print(f"[red]Error: {result.error}[/red]")
-            raise typer.Exit(1)
-
-        # Display success message using formatter
-        new_scene_id: SceneIdentifier = SceneIdentifier(
-            project=project,
-            scene_number=(
-                reference_scene + 1 if position == "after" else reference_scene
-            ),
-            season=season,
-            episode=episode,
-        )
-        formatter.format_operation_result(
-            "add",
-            True,
-            new_scene_id,
-            details={"renumbered_scenes": result.renumbered_scenes},
-        )
-
-    except Exception as e:
-        logger.error(f"Failed to add scene: {e}")
-        console.print(f"[red]Failed to add scene: {e}[/red]")
-        raise typer.Exit(1) from e
+    else:
+        scene_num = result.created_scene.number if result.created_scene else "unknown"
+        handler.handle_success(f"Scene added successfully as scene {scene_num}")
 
 
 @scene_app.command(name="update")
-def update_scene(
+@async_cli_command
+async def update_scene(
     project: Annotated[
         str, typer.Option("--project", "-p", help="Project/script name")
     ],
     scene: Annotated[int, typer.Option("--scene", "-s", help="Scene number")],
-    safe: Annotated[
-        bool, typer.Option("--safe", help="Check for conflicts before updating")
-    ] = False,
-    last_read: Annotated[
-        str | None,
-        typer.Option(
-            "--last-read",
-            help="ISO timestamp of when scene was last read (for --safe mode)",
-        ),
-    ] = None,
     season: Annotated[
         int | None, typer.Option("--season", help="Season number (for TV)")
     ] = None,
@@ -286,6 +245,14 @@ def update_scene(
         str | None,
         typer.Option("--content", help="New scene content (or pipe from stdin)"),
     ] = None,
+    check_conflicts: Annotated[
+        bool,
+        typer.Option(
+            "--check-conflicts",
+            help="Check if scene was modified since last read",
+        ),
+    ] = False,
+    json_output: Annotated[bool, typer.Option("--json", help="Output as JSON")] = False,
     config: Annotated[
         Path | None,
         typer.Option(
@@ -295,95 +262,64 @@ def update_scene(
         ),
     ] = None,
 ) -> None:
-    r"""Update a scene with optional conflict checking.
+    """Update an existing scene.
 
-    By default, updates happen immediately without conflict checking.
-    Use --safe flag for conflict detection (requires --last-read timestamp).
-    Content must be valid Fountain format.
+    Content can be provided via --content or piped from stdin.
 
     Examples:
-        # Simple update (no conflict checking)
-        scriptrag scene update --project "inception" --scene 42
-
-        # Safe update (with conflict checking)
-        scriptrag scene update --safe --project "inception" --scene 42 \
-            --last-read "2024-01-15T10:30:00"
-
-    Use --content flag or pipe content via stdin.
+        scriptrag scene update --project "inception" --scene 42 --content "..."
+        echo "New content" | scriptrag scene update --project "inception" --scene 42
     """
-    try:
-        # Get content from stdin if not provided
-        if content is None:
-            if sys.stdin.isatty():
-                console.print(
-                    "[red]Error: No content provided. Use --content or "
-                    "pipe from stdin[/red]"
-                )
-                raise typer.Exit(1)
-            content = sys.stdin.read()
+    # Initialize handlers
+    handler = CLIHandler(console)
+    formatter = SceneFormatter(console)
 
-        # Create scene identifier
-        scene_id: SceneIdentifier = SceneIdentifier(
-            project=project,
-            scene_number=scene,
-            season=season,
-            episode=episode,
-        )
+    # Validate inputs
+    project_validator = ProjectValidator()
+    project = project_validator.validate(project)
 
-        # Parse last_read timestamp if provided
-        last_read_dt: datetime | None = None
-        if safe:
-            if not last_read:
-                console.print(
-                    "[red]Error: --last-read timestamp required "
-                    "when using --safe flag[/red]"
-                )
-                raise typer.Exit(1)
-            try:
-                last_read_dt = datetime.fromisoformat(last_read)
-            except ValueError as e:
-                console.print(
-                    f"[red]Error: Invalid timestamp format: {last_read}[/red]"
-                )
-                console.print("[dim]Use ISO format: YYYY-MM-DDTHH:MM:SS[/dim]")
-                raise typer.Exit(1) from e
+    scene_validator = SceneValidator()
+    scene_id = scene_validator.validate(
+        {
+            "project": project,
+            "scene_number": scene,
+            "season": season,
+            "episode": episode,
+        }
+    )
 
-        # Load settings with proper precedence
-        settings = load_config_with_validation(config)
+    # Get content
+    if content is None:
+        content = handler.read_stdin(required=True)
 
-        # Initialize API
-        api: SceneManagementAPI = SceneManagementAPI(settings=settings)
+    # Validate content
+    if content is not None:
+        content_validator = SceneContentValidator()
+        content = content_validator.validate(content)
+    else:
+        handler.handle_error(ValueError("No content provided"), json_output)
+        return
 
-        # Update scene
-        from scriptrag.api.scene_models import UpdateSceneResult
+    # Load settings
+    settings = load_config_with_validation(config)
 
-        result: UpdateSceneResult = asyncio.run(
-            api.update_scene(
-                scene_id, content, check_conflicts=safe, last_read=last_read_dt
-            )
-        )
+    # Initialize API
+    api = SceneManagementAPI(settings=settings)
 
-        if not result.success:
-            console.print(f"[red]Error: {result.error}[/red]")
-            if result.validation_errors:
-                formatter.format_validation_errors(result.validation_errors)
-            raise typer.Exit(1)
+    # Update scene through API
+    result = await api.update_scene(scene_id, content, check_conflicts=check_conflicts)
 
-        # Display success message using formatter
-        formatter.format_operation_result(
-            "update",
-            True,
-            scene_id,
-        )
+    if not result.success:
+        handler.handle_error(Exception(result.error), json_output)
 
-    except Exception as e:
-        logger.error(f"Failed to update scene: {e}")
-        console.print(f"[red]Failed to update scene: {e}[/red]")
-        raise typer.Exit(1) from e
+    # Format and output result
+    output_format = OutputFormat.JSON if json_output else OutputFormat.TEXT
+    formatter.print(result, output_format)
 
 
 @scene_app.command(name="delete")
-def delete_scene(
+@async_cli_command
+async def delete_scene(
     project: Annotated[
         str, typer.Option("--project", "-p", help="Project/script name")
     ],
@@ -394,9 +330,10 @@ def delete_scene(
     episode: Annotated[
         int | None, typer.Option("--episode", "-e", help="Episode number (for TV)")
     ] = None,
-    confirm: Annotated[
-        bool, typer.Option("--confirm", help="Confirm deletion")
+    force: Annotated[
+        bool, typer.Option("--force", "-f", help="Skip confirmation prompt")
     ] = False,
+    json_output: Annotated[bool, typer.Option("--json", help="Output as JSON")] = False,
     config: Annotated[
         Path | None,
         typer.Option(
@@ -406,56 +343,59 @@ def delete_scene(
         ),
     ] = None,
 ) -> None:
-    r"""Delete a scene with automatic renumbering.
-
-    Requires --confirm flag to prevent accidental deletions.
-    Automatically renumbers subsequent scenes.
+    """Delete a scene with automatic renumbering.
 
     Examples:
-        scriptrag scene delete --project "inception" --scene 42 --confirm
+        scriptrag scene delete --project "inception" --scene 42 --force
     """
-    if not confirm:
-        console.print(
-            "[yellow]Warning: This will permanently delete the scene.[/yellow]"
+    # Initialize handlers
+    handler = CLIHandler(console)
+
+    # Validate inputs
+    project_validator = ProjectValidator()
+    project = project_validator.validate(project)
+
+    scene_validator = SceneValidator()
+    scene_id = scene_validator.validate(
+        {
+            "project": project,
+            "scene_number": scene,
+            "season": season,
+            "episode": episode,
+        }
+    )
+
+    # Confirm deletion if not forced
+    if not force and not json_output:
+        confirm = typer.confirm(f"Delete scene {scene_id.key}?")
+        if not confirm:
+            handler.handle_success("Deletion cancelled")
+            return
+
+    # Load settings
+    settings = load_config_with_validation(config)
+
+    # Initialize API
+    api = SceneManagementAPI(settings=settings)
+
+    # Delete scene through API
+    result = await api.delete_scene(scene_id)
+
+    if not result.success:
+        handler.handle_error(Exception(result.error), json_output)
+
+    # Format success message
+    if json_output:
+        handler.handle_success(
+            f"Scene {scene_id.key} deleted successfully",
+            {
+                "deleted_scene": scene_id.key,
+                "renumbered_count": len(result.renumbered_scenes),
+            },
+            json_output=True,
         )
-        console.print("Add --confirm flag to proceed with deletion.")
-        raise typer.Exit(0)
-
-    try:
-        # Create scene identifier
-        scene_id: SceneIdentifier = SceneIdentifier(
-            project=project,
-            scene_number=scene,
-            season=season,
-            episode=episode,
-        )
-
-        # Load settings with proper precedence
-        settings = load_config_with_validation(config)
-
-        # Initialize API
-        api: SceneManagementAPI = SceneManagementAPI(settings=settings)
-
-        # Delete scene
-        from scriptrag.api.scene_models import DeleteSceneResult
-
-        result: DeleteSceneResult = asyncio.run(
-            api.delete_scene(scene_id, confirm=True)
-        )
-
-        if not result.success:
-            console.print(f"[red]Error: {result.error}[/red]")
-            raise typer.Exit(1)
-
-        # Display success message using formatter
-        formatter.format_operation_result(
-            "delete",
-            True,
-            scene_id,
-            details={"renumbered_scenes": result.renumbered_scenes},
-        )
-
-    except Exception as e:
-        logger.error(f"Failed to delete scene: {e}")
-        console.print(f"[red]Failed to delete scene: {e}[/red]")
-        raise typer.Exit(1) from e
+    else:
+        msg = f"Scene {scene_id.key} deleted successfully"
+        if result.renumbered_scenes:
+            msg += f" ({len(result.renumbered_scenes)} scenes renumbered)"
+        handler.handle_success(msg)
